Patch 12.2 is a minor update for Riot Games Servers

+++ R3nzSkin/offsets.hpp
@@ -12,31 +12,31 @@ namespace offsets {
 	};
 
 	namespace global {
-		inline std::uintptr_t Player{ 0 };
-		inline std::uintptr_t ChampionManager{ 0 };
-		inline std::uintptr_t Riot__g_window{ 0 };
-		inline std::uintptr_t ManagerTemplate_AIMinionClient_{ 0 };
-		inline std::uintptr_t ManagerTemplate_AIHero_{ 0 };
-		inline std::uintptr_t GameClient{ 0 };
+		inline std::uint32_t Player{ 0 };
+		inline std::uint32_t ChampionManager{ 0 };
+		inline std::uint32_t Riot__g_window{ 0 };
+		inline std::uint32_t ManagerTemplate_AIMinionClient_{ 0 };
+		inline std::uint32_t ManagerTemplate_AIHero_{ 0 };
+		inline std::uint32_t GameClient{ 0 };
 	};
 
 	namespace AIBaseCommon {
-		inline std::uintptr_t CharacterDataStack{ 0 };
-		inline std::uintptr_t SkinId{ 0 };
+		inline std::uint32_t CharacterDataStack{ 0 };
+		inline std::uint32_t SkinId{ 0 };
 	};
 
 	namespace MaterialRegistry {
-		inline std::uintptr_t D3DDevice{ 0 };
+		inline std::uint32_t D3DDevice{ 0 };
 #ifdef _RIOT
-		inline std::uintptr_t SwapChain{ 0 };
+		inline std::uint32_t SwapChain{ 0 };
 #endif
 	};
 
 	namespace functions {
-		inline std::uintptr_t Riot__Renderer__MaterialRegistry__GetSingletonPtr{ 0 };
-		inline std::uintptr_t translateString_UNSAFE_DONOTUSE{ 0 };
-		inline std::uintptr_t CharacterDataStack__Push{ 0 };
-		inline std::uintptr_t CharacterDataStack__Update{ 0 };
-		inline std::uintptr_t GetGoldRedirectTarget{ 0 };
+		inline std::uint32_t Riot__Renderer__MaterialRegistry__GetSingletonPtr{ 0 };
+		inline std::uint32_t translateString_UNSAFE_DONOTUSE{ 0 };
+		inline std::uint32_t CharacterDataStack__Push{ 0 };
+		inline std::uint32_t CharacterDataStack__Update{ 0 };
+		inline std::uint32_t GetGoldRedirectTarget{ 0 };
 	};
 };
